requirements.txt
fastapi~=0.115.8
PyGithub~=2.5.0
pandas~=2.2.3
numpy~=2.2.2